(MISC): Tidied `RustCrateUtil`;         Moved `RustCrateUtil` `toolchain.*` -> into `util.*`
