+++ src/test/kotlin/org/rust/cargo/toolchain/CargoProjectResolveTestCase.kt
@@ -9,6 +9,7 @@ import com.intellij.util.ui.UIUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.rust.cargo.RustWithToolchainTestCaseBase
 import org.rust.cargo.projectSettings.toolchain
+import org.rust.cargo.util.cargoProject
 import org.rust.cargo.util.getService
 
 class CargoProjectResolveTestCase : RustWithToolchainTestCaseBase() {
